chore(agent): satisfy daemon go lint
